features/requirements.txt
Appium-Python-Client==0.20
behave==1.2.5
nose==1.3.7
parse==1.6.6
parse-type==0.3.4
selenium==2.48.0
six==1.10.0
wheel==0.24.0
